requirements.txt
Django==2.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
pkg-resources==0.0.0
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
